requirements.txt
# Core Dependencies
pymongo>=4.0.0
neo4j>=5.0.0
openai>=1.0.0
streamlit>=1.28.0

# Data Processing
pandas>=1.5.0
numpy>=1.21.0
lxml>=4.9.0

# Visualization
plotly>=5.15.0
networkx>=3.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.28.0

# Jupyter Support
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional Libraries
tqdm>=4.64.0
scikit-learn>=1.2.0
